+++ project_2/src/main/java/com/revature/repository/entity/FriendRequest.java
@@ -14,44 +14,44 @@ public class FriendRequest {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int request_id;
+	private long request_id;
 	
-	private int sender_id;
+	private long sender_id;
 	
-	private int receiver_id;
+	private long receiver_id;
 
 	public FriendRequest() {
 		super();
 	}
 
-	public FriendRequest(int request_id, int sender_id, int receiver_id) {
+	public FriendRequest(long request_id, long sender_id, long receiver_id) {
 		super();
 		this.request_id = request_id;
 		this.sender_id = sender_id;
 		this.receiver_id = receiver_id;
 	}
 
-	public int getRequest_id() {
+	public long getRequest_id() {
 		return request_id;
 	}
 
-	public void setRequest_id(int request_id) {
+	public void setRequest_id(long request_id) {
 		this.request_id = request_id;
 	}
 
-	public int getSender_id() {
+	public long getSender_id() {
 		return sender_id;
 	}
 
-	public void setSender_id(int sender_id) {
+	public void setSender_id(long sender_id) {
 		this.sender_id = sender_id;
 	}
 
-	public int getReceiver_id() {
+	public long getReceiver_id() {
 		return receiver_id;
 	}
 
-	public void setReceiver_id(int receiver_id) {
+	public void setReceiver_id(long receiver_id) {
 		this.receiver_id = receiver_id;
 	}
